api/requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
frozendict==2.3.10
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moralis==0.1.49
paho-mqtt==1.6.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
